+++ tests/unit/test_validate_ind_cqc_ascwds_and_pir_data.py
@@ -2,7 +2,7 @@ import unittest
 
 from unittest.mock import Mock, patch
 
-import jobs.validate_estimated_missing_ascwds_filled_posts_data as job
+import jobs.validate_impute_ind_cqc_ascwds_and_pir_data as job
 
 from tests.test_file_data import ValidateEstimatedMissingAscwdsFilledPostsData as Data
 from tests.test_file_schemas import (
